PR: Frontend mid-project housekeeping #66  (#67)
See subtasks and progress in issue description:
https://github.com/zpi-2023/senso-frontend/issues/66

The diff is a mess as always. There are no changes to logic though. What
I did is enable a lot of lints and then go over and fix any issues
found. And some edge cases found were actual bugs in our code, so that's
great. Easy-to-make errors should now be prevented and there are also
some conveniences thrown in.